+++ Assets/Scripts/CharacterSelect/CharacterSelectView.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using DG.Tweening;
 using UnityEngine;
 using UnityEngine.UI;
+using Utilities;
 
 namespace CharacterSelect
 {
