+++ Assets/Scripts/MainMenu/MainMenu.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
+using Utilities;
 
 namespace MainMenu
 {
